requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
gmail==0.6.3
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
utils==1.0.1
